Initial commit
Chrome extension to show how many hours are left to work for today to
complete the quoata of hours for month.

Baseline - Get the page source into popup.html.

As of now the source code of the page that we are in, will be sent to
popup.html.

Next step is to parse and filter that HTML information into following
things : 1. Number of hours left for today 2. May be graph (optional)

Future plan : If the user has not logged in login with credentials and
get the page source of that specific page.